+++ lib/src/extensions/date_time.dart
@@ -1,11 +1,13 @@
 import 'package:nyxx_extensions/src/utils/formatters.dart';
 
-extension TimestampStyleDateTime on DateTime {
+/// Extensions on [DateTime].
+extension DateTimeExtensions on DateTime {
   /// Formats the [DateTime] into a date string timestamp.
   String format([TimestampStyle style = TimestampStyle.none]) => formatDate(this, style);
 }
 
-extension TimestampStyleDuration on Duration {
+/// Extensions on [Duration].
+extension DurationExtensions on Duration {
   /// Formats the [Duration] into a date string timestamp.
   /// The style will always be relative to represent as a duration on Discord.
   String format() => formatDate(DateTime.now().add(this), TimestampStyle.relativeTime);